Initial checkin of VS integration effort.
Doc on prerequesites in README.md

Based on experiments, the tsconfig.json outDir option confuses Visual Studios integration with Mocha, so I'm removing that option.   This means the .js and .js.map files will be created in the same directory as the .ts files that they are based on.

Added some known Visual Studio .gitignore terms.