Move Decorator functions to their own module
Lots of files touched, but with the exception of Stubs.ts and Decorators.ts, it is all very minor and mechanical.

I was starting to get circular dependency problems involving `Stubs.ts`.   With the decorator functions in that module being very low level, while other true stubs are much higher level, this repackaging seems the best solution.