Eliminate `children` from Tree, RuleNode, etc
The basic idea here is that our Tree interface now inherits from ArrayLike<Tree>, so that `.length` and access to children via `[n]` are uncomplicated.

To implement this, `TerminalNode` gets a dummy implementation without actually being an array, while the `RuleNode` extends `Array<RuleNode|TerminalNode>`

Includes fixup as needed to eliminate references to `children` as a property.