Update build process for speed.  Operates across packages in this repository
To run antlr, we use the `lerna run antlr` command.   Any package that has an antlr script is run, lerna controls the ordering based on dependencies declared in the package.json files.

The typescript compiler has multi-package build-in (in V3+).  Combined
with a new incremetal compile feature, this is a big time savings.